Squashed commit of the following:
commit a841de5ab3ba3111e714e71056b098c891542bc7
Author: Mark Otway <mark@otway.com>
Date:   Sat Aug 15 10:42:29 2026 +0100

    3rd time lucky

commit b2ff859d88673acf0373ecd08c4b2eda181e3f7d
Author: Mark Otway <mark@otway.com>
Date:   Sat Aug 15 10:30:32 2026 +0100

    Make it build

commit ca0d8363b51fa1ecd6c00c2929f433d5bddbc9e1
Author: Mark Otway <mark@otway.com>
Date:   Sat Aug 15 09:39:01 2026 +0100

    Use decimal for batteryCapacitySoc

commit 49df4f7140db4e4708fc94bdfde0c633d9e4c5d4
Author: Mark Otway <mark@otway.com>
Date:   Fri Aug 14 16:01:54 2026 +0100

    Handle multiline log entries in LogView

commit 427846520af5e2e572764cec65fadaf3d968cc7d
Author: Mark Otway <mark@otway.com>
Date:   Thu Jul 30 08:41:39 2026 +0100

    More logging

commit 19c5844d8e12bd1ec23895d3363185d6e1b95994
Author: Mark Otway <mark@otway.com>
Date:   Thu Jul 30 07:23:45 2026 +0100

    Logging

commit ae333caa5334af5249c91c018fcd8ec1bdd9e8ad
Merge: 961f051 d42f4a0
Author: Mark Otway <mark@otway.com>
Date:   Wed Jul 29 09:29:40 2026 +0100

    Merge branch 'main' into develop

commit 961f051c3093f3f7c4f86cc60d2c396d90e6b996
Author: Mark Otway <mark@otway.com>
Date:   Mon Jul 27 12:00:28 2026 +0100

    #303 Try and fix

commit 886945fc54602665dfc6fe3a089b9762a73bc5c6
Author: Mark Otway <mark@otway.com>
Date:   Mon Jul 27 11:55:28 2026 +0100

    Possible fix for Tom

commit 0f11e44846e0a8025ca770a19b3a9eb6951604da
Author: Mark Otway <mark@otway.com>
Date:   Mon Jul 13 09:31:01 2026 +0100

    Project 2 days for IOG SMB rates

commit 6f10dd662deb0cd1a0a2e4e7ebdb57de7eb7fd1f
Author: Mark Otway <mark@otway.com>
Date:   Fri Jul 10 20:44:47 2026 +0100

    Revert revert

commit 0c3dcfbc0a8af1b39fe812c6be6be725b364f673
Author: Mark Otway <mark@otway.com>
Date:   Fri Jul 10 20:44:34 2026 +0100

    Revert

commit ed30bec23b1d84dbc81e11a6c0456942e99c4650
Author: Mark Otway <mark@otway.com>
Date:   Fri Jul 10 20:39:53 2026 +0100

    Update OctopusAPI.cs

commit df796f5d08587921d4704272980219196a524c9b
Author: Mark Otway <mark@otway.com>
Date:   Fri Jul 10 19:38:29 2026 +0100

    Fix missing IOG slots

commit 9f7054ccb498b74c5e16776bd31c6587d5871c80
Author: Mark Otway <mark@otway.com>
Date:   Fri Jul 10 10:35:53 2026 +0100

    Remove hard-coding

commit 8d91323a36b06d1cbfb6f6de24d4b8a1b390f5e2
Author: Mark Otway <mark@otway.com>
Date:   Fri Jul 10 10:32:42 2026 +0100

    New method for getting rates

commit 9fceb00388ab6c7683f110ffb139726bca45381d
Author: Mark Otway <mark@otway.com>
Date:   Thu Jul 9 22:47:55 2026 +0100

    Tidying

commit 05074d264f447bd25f1a133e280d151deda6b2df
Author: Mark Otway <mark@otway.com>
Date:   Thu Jul 9 09:09:21 2026 +0100

    Clear Override when checkbox unchecked

commit 5a74b08e91ce96dc889eead018ab810e7606c5bf
Author: Mark Otway <mark@otway.com>
Date:   Thu Jul 9 08:53:35 2026 +0100

    Oops

commit faeac5976326dd72edaa0d3c409f267cb5dd0e27
Author: Mark Otway <mark@otway.com>
Date:   Thu Jul 9 08:22:24 2026 +0100

    Config Settings Rework

commit 9b72df68ed051f87787d6f5c2a4ec7b7abc80eba
Author: Mark Otway <mark@otway.com>
Date:   Wed Jul 8 16:42:48 2026 +0100

    Update OctopusAPI.cs

commit 8364a33206af47290e8f5a41088664c5b69da386
Author: Mark Otway <mark@otway.com>
Date:   Wed Jul 8 09:48:19 2026 +0100

    #307 Support new tariff code

commit 9981d5f387b088d9bee90ef7371c0b0b742b177d
Author: Mark Otway <mark@otway.com>
Date:   Fri Jul 3 11:28:14 2026 +0100

    Pre-charge can be a boost charge

commit b248123fad9448a45bd56e04f71dd3bad543b6aa
Author: Mark Otway <mark@otway.com>
Date:   Tue Jun 30 11:57:55 2026 +0100

    Global UI consts

+++ SolisManager.Shared/Models/HistoryEntry.cs
@@ -9,7 +9,7 @@ public class HistoryEntry
     public DateTime Start { get; set; }
     public DateTime End { get; set; }
     public decimal Price { get; set; }
-    public int BatterySOC { get; set; }
+    public decimal BatterySOC { get; set; }
     public PriceType Type { get; set; }
     public SlotAction Action { get; set; }
     public string Reason { get; set; } = string.Empty;
